Move preview components to library and create a DesignLibraryFragment for applications to use
