+++ components/build.gradle.kts
@@ -49,10 +49,13 @@ mavenPublishing {
 }
 
 dependencies {
+    implementation(libs.androidx.activity.compose)
+    implementation(libs.androidx.fragment)
     implementation(libs.androidx.ui)
     implementation(libs.androidx.ui.graphics)
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
+    implementation(libs.kotlin.reflect)
     api(libs.green.tokens)
     debugImplementation(libs.androidx.ui.tooling)
 }
