+++ components/src/main/kotlin/se/seb/gds/views/GdsButtonView.kt
@@ -3,8 +3,8 @@ package se.seb.gds.views
 import android.content.Context
 import android.util.AttributeSet
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.platform.AbstractComposeView
 import se.seb.gds.atoms.GdsButton
