--- components/src/test/java/se/seb/gds/components/ExampleUnitTest.kt
@@ -1,9 +1,0 @@
-package se.seb.gds.components
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-}
